requirements.txt
aiohttp
django
django-q
django-q2
djangorestframework
drf-yasg
pre-commit
python-dotenv
requests
